Set the licensing. Change README. Add images.
Added a link to microscope's official website. Changed the licensing information to CC Attribution, and a blurb establishing that adding to the repository (via pull request) is consent to publish using the same license. That way we can all relax about sharing it or using it for our own nefarious purposes.

Added a link to the Slack.